--- src/vector_diagram/AddButtonListener.java
@@ -1,17 +1,0 @@
-package vector_diagram;
-
-import javax.swing.*;
-import java.awt.event.*;
-import java.awt.*;
-
-public class AddButtonListener implements ActionListener {
-	private JFrame frame;
-	
-	public AddButtonListener() {
-		
-	}
-	
-	public void actionPerformed(ActionEvent e) {
-		
-	}
-}
