+++ src/vector_diagram/Vector_diagram.java
@@ -11,6 +11,9 @@ import javax.swing.GroupLayout;
 import javax.swing.GroupLayout.Alignment;
 import javax.swing.JTabbedPane;
 import javax.swing.border.TitledBorder;
+
+import listener.AddButtonListener;
+
 import javax.swing.JButton;
 import java.awt.event.ActionListener;
 import java.awt.event.ActionEvent;
@@ -29,8 +32,7 @@ public class Vector_diagram extends JFrame implements Runnable {
 
 	private static final long serialVersionUID = -7525251402309591219L;
 	private JPanel mainPane;
-	private JTextField referenceTextField;
-
+	
 	@Override
 	public void run() {
 
@@ -99,61 +101,7 @@ public class Vector_diagram extends JFrame implements Runnable {
 						.addComponent(addButton))
 					.addContainerGap())
 		);
-
-		JPanel panel = new JPanel();
-		components.addTab("", null, panel, null);
-
-		JLabel componentLabel = new JLabel("Component");
-
-		JComboBox componentBox = new JComboBox();
-		componentBox.setToolTipText("Component Type");
-		
-		JLabel referenceLabel = new JLabel("Reference");
-		
-		referenceTextField = new JTextField();
-		referenceTextField.setToolTipText("e.g. R1, R2, L1");
-		referenceTextField.setColumns(10);
-		
-		JLabel valueLabel = new JLabel("Value");
-		
-		JSpinner valueSpinner = new JSpinner();
-		valueSpinner.setModel(new SpinnerNumberModel(new Double(0), new Double(0), null, new Double(1)));
-		GroupLayout gl_panel = new GroupLayout(panel);
-		gl_panel.setHorizontalGroup(
-			gl_panel.createParallelGroup(Alignment.LEADING)
-				.addGroup(gl_panel.createSequentialGroup()
-					.addContainerGap()
-					.addGroup(gl_panel.createParallelGroup(Alignment.LEADING)
-						.addComponent(componentLabel)
-						.addComponent(valueLabel)
-						.addComponent(referenceLabel))
-					.addGap(18)
-					.addGroup(gl_panel.createParallelGroup(Alignment.TRAILING)
-						.addGroup(gl_panel.createParallelGroup(Alignment.LEADING, false)
-							.addComponent(valueSpinner)
-							.addComponent(referenceTextField, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE))
-						.addComponent(componentBox, GroupLayout.PREFERRED_SIZE, 86, GroupLayout.PREFERRED_SIZE))
-					.addContainerGap(408, Short.MAX_VALUE))
-		);
-		gl_panel.setVerticalGroup(
-			gl_panel.createParallelGroup(Alignment.LEADING)
-				.addGroup(gl_panel.createSequentialGroup()
-					.addContainerGap()
-					.addGroup(gl_panel.createParallelGroup(Alignment.LEADING)
-						.addGroup(gl_panel.createSequentialGroup()
-							.addComponent(componentBox, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
-							.addPreferredGap(ComponentPlacement.RELATED)
-							.addGroup(gl_panel.createParallelGroup(Alignment.BASELINE)
-								.addComponent(referenceTextField, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
-								.addComponent(referenceLabel))
-							.addPreferredGap(ComponentPlacement.RELATED)
-							.addGroup(gl_panel.createParallelGroup(Alignment.BASELINE)
-								.addComponent(valueSpinner, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
-								.addComponent(valueLabel)))
-						.addComponent(componentLabel))
-					.addContainerGap(166, Short.MAX_VALUE))
-		);
-		panel.setLayout(gl_panel);
 		mainPane.setLayout(gl_mainPane);
+		addButton.addActionListener(new AddButtonListener(components));
 	}
 }
